Fashion boutique
A sequence of clothes in a box is given. The task is to arrange the clothes in racks depending on their value. When the sum value of the clothes exceeds the capacity of the rack, we go to the next rack. The task is to calculate how many racks are needed to store all the clothes.