Truck tour
There is a circle road with N petrol pumps. The truck consumes 1 liter of petrol per 1 kilometer, and its tank has infinite petrol capacity. We receive as input the amount of petrol the petrol pump will give us and the distance from a petrol pump to the next one. The task is to calculate the first petrol pump from where the truck will be able to complete the circle.