Diagonal difference
The program reads a number representing rows and cols of a matrix. Each row consists of ASCII characters. After that, a symbol is input. The program finds the first occurrence of the symbol in the matrix and prints its position. If there is no such symbol it prints an error message.