Supermarket
This code reads names and adds them to a queue until "End" is received. If "Paid" is received, it prints every customer name and empties the queue, otherwise, a client is received and it's added to the queue. When "End" is received, it prints the count of the remaining people in the queue.