Fill the matrix
The program receives a number N - the dimensions of a square matrix and a method (A or B) to fill out the matrix. 
A is filling it top to bottom column by column and B is filling it top to bottom and then bottom to top again column by column.