Counting words
Given a string paragraph, the task is to count the number of different words, and the number of punctuation marks. The output should be alphabetically sorted (based on the ASCII table). 

Input:
You are given a paragraph from a book. The paragraph can
have words, separated by one or more spaces, commas,
points, question marks ? and exclamation marks !.