Tournament
The input is a string of team names, then the team matches with their results. The winning team gets 3 points and when the game is draw, both teams get 1 point. The output should be: 
TeamName - TotalPoints - TotalScoredGoals - TotalReceivedGoals
At the end, all the teams with the highest points are printed.