Find
The program reads info about companies in the format: {name} {id} until "end" is received. After, it reads a search id and finds the company with that id. 
If there is such a company, the program prints the info about it. If no such company is found, "[not found]" is printed.