Try parse
The program reads two strings, each of which is a valid integer, and parses them into integer numbers. 
If both numbers can be parsed, the program calculates and prints the sum. Otherwise, two lines are printed:
One of them for the first number, the other for the second number 
Each of those lines contains either the input string for that number, but if the number was not successfully parsed, the line starts with "[error] " followed by the input string for the number.